+++ src/components/blocks/QuestionCards/QuestionCards.tsx
@@ -15,8 +15,8 @@ import { isEmpty } from 'lodash';
 import { useTranslation } from 'react-i18next';
 import useSWR, { mutate } from 'swr';
 
+import { getQuestionsEndpoint } from '../../../api/endpoints';
 import { archieveQuestion } from '../../../api/questions';
-import { getQuestionsEndpoint } from '../../../features/swr/index';
 import {
   Categories,
   CategoryControl,
