+++ src/components/views/Login.tsx
@@ -14,10 +14,7 @@ import { useTranslation } from 'react-i18next';
 import useSWR, { mutate } from 'swr';
 
 import { login } from '../../api';
-import {
-  IS_AUTHENTICATED_ENDPOINT,
-  USER_ENDPOINT,
-} from '../../features/swr/index';
+import { IS_AUTHENTICATED_ENDPOINT, USER_ENDPOINT } from '../../api/endpoints';
 import { onFormError, registerInput } from '../../helpers/form';
 import useQueryParam, { useRemoveQueryParam } from '../../hooks/useQueryParam';
 import {
