+++ src/components/views/VideoCall.tsx
@@ -31,18 +31,18 @@ import { useLocation, useNavigate, useParams } from 'react-router-dom';
 import { useTheme } from 'styled-components';
 import useSWR from 'swr';
 
+import {
+  getChatEndpoint,
+  RANDOM_CALL_EXIT_PARAM,
+  RANDOM_CALL_EXIT_VALUE,
+  USER_ENDPOINT,
+} from '../../api/endpoints';
 import { callAgain } from '../../api/livekit';
 import { endRandomCallMatch } from '../../api/randomCalls';
 import {
   useChatInputStore,
   useConnectedCallStore,
 } from '../../features/stores';
-import {
-  getChatEndpoint,
-  RANDOM_CALL_EXIT_PARAM,
-  RANDOM_CALL_EXIT_VALUE,
-  USER_ENDPOINT,
-} from '../../features/swr';
 import useIsBelowBreakpoint from '../../hooks/useIsBelowBreakpoint';
 import useKeyboardShortcut from '../../hooks/useKeyboardShortcut';
 import {
