+++ src/hooks/useUnreadNotificationCount.ts
@@ -1,7 +1,7 @@
 import useSWR from 'swr';
 
+import { UNREAD_NOTIFICATIONS_ENDPOINT } from '../api/endpoints';
 import { fetchNotifications } from '../api/notification';
-import { UNREAD_NOTIFICATIONS_ENDPOINT } from '../features/swr/index';
 
 function useUnreadNotificationCount() {
   const response = useSWR(UNREAD_NOTIFICATIONS_ENDPOINT, fetchNotifications, {
